Logout-icon-logout-added
hooks are connected to the dashboard. Logout icon logout feature has been added
