+++ frontend/src/Pages/Newsui.js
@@ -1,0 +1,18 @@
+import React from 'react'
+import Dashboard from './Dashboardui';
+
+const Newsui = () => {
+
+    let news = (
+        <div>
+            <h1>Deneme</h1>
+        </div>
+    );
+
+  return (
+    <div>
+        <Dashboard  news={news}/>
+    </div>
+  )
+}
+export default Newsui;
